JournalApp modified

+++ JournalApp/src/main/java/com/springbootproject1/JournalApp/controller/JournalEntryController.java
@@ -1,16 +1,16 @@
 package com.springbootproject1.JournalApp.controller;
 
 import com.springbootproject1.JournalApp.entity.JournalEntry;
-import com.springbootproject1.JournalApp.repository.JournalEntryRepo;
 import com.springbootproject1.JournalApp.service.JournalEntryService;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.time.LocalDateTime;
-import java.util.Collections;
 import java.util.List;
-import java.util.Optional;
+
 
 @RestController
 @RequestMapping("journal")
@@ -19,41 +19,60 @@ public class JournalEntryController {
     private JournalEntryService entryService;
 
     @PostMapping
-    public boolean createEntry(@RequestBody JournalEntry entry)
+    public ResponseEntity<?> createEntry(@RequestBody JournalEntry entry)
     {
-        entry.setDate(LocalDateTime.now());
-        entryService.saveEntry(entry);
-        return true;
+        try {
+            entry.setDate(LocalDateTime.now());
+            entryService.saveEntry(entry);
+            return new ResponseEntity<>(HttpStatus.CREATED);
+        }
+        catch(Exception e){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
     }
 
     @GetMapping
-    public List<JournalEntry> getAll()
+    public ResponseEntity<?> getAll()
     {
-        return entryService.getEntries();
+        List<JournalEntry> all=entryService.getEntries();
+        if(all!=null && !all.isEmpty())
+            return new ResponseEntity<>(all, HttpStatus.OK);
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
+
     @GetMapping("id/{id}")
-    public List<JournalEntry> getById(@PathVariable ObjectId id)
+    public ResponseEntity<?> getById(@PathVariable ObjectId id)
     {
-        return Collections.singletonList(entryService.findById(id));
+        JournalEntry entry=entryService.findById(id);
+        if(entry!=null)
+            return new ResponseEntity<>(entryService.findById(id), HttpStatus.OK);
+
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+
     }
 
     @PutMapping("id/{id}")
-    public JournalEntry updateEntry(@PathVariable ObjectId id, @RequestBody JournalEntry entry)
+    public ResponseEntity<?> updateEntry(@PathVariable ObjectId id, @RequestBody JournalEntry entry)
     {
         JournalEntry oldEntry=entryService.findById(id);
         if(oldEntry!=null)
         {
             oldEntry.setTitle((entry.getTitle()!=null && !entry.getTitle().isEmpty())?entry.getTitle(): oldEntry.getTitle());
             oldEntry.setContent((entry.getContent()!=null && !entry.getContent().isBlank())?entry.getContent(): oldEntry.getContent());
+            entryService.saveEntry(oldEntry);
+            return new ResponseEntity<>(oldEntry,HttpStatus.OK);
         }
-        entryService.saveEntry(oldEntry);
-        return oldEntry;
+        return new ResponseEntity<>(HttpStatus.NOT_FOUND);
     }
 
     @DeleteMapping("id/{id}")
-    public boolean deleteEntry(@PathVariable ObjectId id)
+    public ResponseEntity<?> deleteEntry(@PathVariable ObjectId id)
     {
+        JournalEntry entry=entryService.findById(id);
+        if(entry==null)
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+
         entryService.deleteById(id);
-        return true;
+        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 }
